Renamed 'MasterViewController' to 'ListViewController' to reflect more on what the view controller actually displays.
I added a hook for the user selecting annotations (spots located on the map), which can trigger
an event inside of 'MapViewController'. I think this would be useful for bringing in a popoverview
to show more information about the selected business, or bring in the listviewcontroller so the user can
confirm that this is indeed the location that she/he wants to check into.